TSS-43 Removed the private Processor and FieldData endpoints from Aquarius.Client
This will give us the option of redistributing Aquarius.Client.dll to partner integrators without exposing the interal APIs that can and do change from release to release.

1) Removed the generated FieldData and Processor service model classes

2) Removed the FieldData and Processor end point definitions

3) Removed the FieldData and Processor end points from the code generator script

4) Removed the FieldDataClient and ProcessorClient properties from the IAquariusClient interface

5) Added back the `GET /apps/v1/version` endpoint just to enable the version detector logic.

We don't need all of the FieldData service model to implement this. Just the one REST API method and response DTO will do.

6) Added the ClientHelper.CloneAuthenticatedClient() method and unit tests

7) Added the IAquariusClient.RegisterCustomClient() method to allow WebPortal to register the private FieldData and Processor APIs

8) Added the CloneAuthenticatedClient() and CloneAuthenticatedClientWithOverrideMethod() helper methods to IAquariusClient

This will allow consumers (like WebPortal) to use nothing but interfaces to integrate with NG and remain easily unit-testable.