TSS-57 Added app.config overrides for AquariusSystemDetector
The app.config/web.config can now specify an optional app setting to specify a list of hostname=fakeVersion pairs.

Any hostname matching this list will not be version probed, but instead will be treated as if the fakeVersion response was returned.

This hack will be used to keep a WSASK production system compatible with WebPortal. Yuck!