PF-27 Added support for anonymous POST requests to Samples
There are a few API POST requests that don't fully follow RequestDTO/ResponseDTO pattern.
These requests accept anonymous lists of inputs, and return a ResponseDTO.

So we expose the InvokeWebServiceMethod() methods on the interface, so that the SamplesConnector can execute the lower-level JsonServiceClient methods with a correct SamplesContext.

This avoids polluting the longer-term .NET SDK interface with requests that are more brittle.

So we need to expose the less wrist-friendly Verb<TResponse>(string relativeOrAbsoluteUri, object request) methods until the Samples APIs are improved.